data_processing/requirements.txt
pandas
xlrd
openpyxl
sqlalchemy
psycopg2
python-dotenv
